Put load_config in the base_simulation file
